navigate: Move scroll_to_selected to message_viewport.
